ENH: Provides a model for an ITK style filter
We anticipate that ITK will be heavily used in our project. Some of our
classes may need to be a part of the ITK pipline. This basic example provides
a model that may be used when writing classes that conform to the ITK standard
and use Smart Pointers ->SetInput ->GetOutput ->Update functionality etc. and
can be used within the ITK pipeline.