ENH: Adds a basic "proof of runnability" test
A basic "does it run" test is added to the BRAINSDeface module. It
checks to see if the program can be run. Our current testing data is not public
and so cannot be hosted on midas. Instead it is being put into:
/scratch/testData/BRAINSDefaceData
We should add a cmake argument that can change this directory so its specified
at configure time.