ENH: CTest checks if brain data has been altered
Uses the brain label map to mask the brain of the original image and defaced image.
Then it uses ITK absolute value difference image filter to create a difference image.
Then all the values of the stistics image is summed together, and the tests fails if the sum is not 0.