ENH: Allows a face mask to be created from landmarks
Using the Left eye, Right eye and dens_axis landmarks, a mask is created
that covers only facial features and not the brain. This is useful, as some
images will not come with an atlas, just landmarks. Currently this is still
in "proof of concept" phase. It needs to be cleaned up, put into a separate file
and tests need to be added.