ENH: Places landmark mask funtion into different file
For some reason, the program doesn't work with slicer, but it does work
from the command line. It could be that my slicer setup is wrong. This could
also use an option to choose using the landmarks or using the label atals, or seeing what is available and choosing automatically.